client & backend fixes
