+++ client/src/app/main.tsx
@@ -1,6 +1,6 @@
 import { createRoot } from "react-dom/client";
 import { App } from "./App.tsx";
-import { LoginPage } from "@/pages/index.ts";
+import { LandingPage } from "@/pages/index.ts";
 
 createRoot(document.getElementById("root")!).render(
   <>
