+++ client/src/app/router/index.tsx
@@ -1,13 +1,12 @@
 import { createBrowserRouter } from "react-router-dom";
-// import { Layout } from "./Layout/Layout";
-import { AuthPage, HomePage, RegPage, NotFound, LoginPage } from "@/pages";
+import {HomePage, LandingPage, NotFound} from "@/pages";
 import { UserItem } from "@/entities/user";
+import {Layout} from "@/app/router/Layout/Layout.tsx";
+import {LandingLayout} from "@/app/router/Layout/LandingLayout.tsx";
 
 export enum CLIENT_ROUTES {
+  LANDING = "/welcome",
   HOME = "/",
-  AUTH = "/authorization",
-  REG = "/registration",
-  LOGIN = "/login",
   NOT_FOUND = "*",
   PROFILE = "/user/:id",
 }
@@ -26,20 +25,18 @@ export const router = createBrowserRouter([
         element: <UserItem />,
       },
       {
-        path: CLIENT_ROUTES.AUTH,
-        element: <AuthPage />,
-      },
-      {
-        path: CLIENT_ROUTES.REG,
-        element: <RegPage />,
-      },
-      {
         path: CLIENT_ROUTES.NOT_FOUND,
         element: <NotFound />,
       },
+    ],
+  },
+  {
+    path: CLIENT_ROUTES.LANDING,
+    element: <LandingLayout />,
+    children: [
       {
-        path: CLIENT_ROUTES.LOGIN,
-        element: <LoginPage />,
+        path: CLIENT_ROUTES.LANDING,
+        element: <LandingPage />,
       },
     ],
   },
