+++ client/src/features/ticket/CreateTicketButton/ui/CreateTicketButton.tsx
@@ -1,0 +1,16 @@
+
+
+export function CreateTicketButton() {
+    return (
+        <>
+            <div className="level">
+                <div className="level-left">
+                </div>
+                <div className="level-right">
+                    <button className="button is-info">Создать задачу</button>
+                </div>
+            </div>
+        </>
+    );
+}
+
