--- client/src/pages/AuthPage/index.ts
@@ -1,1 +1,0 @@
-// export { AuthPage } from './ui/AuthPage';
