--- client/src/pages/AuthPage/ui/AuthPage.tsx
@@ -1,9 +1,0 @@
-// import { AuthorizationForm } from "@/features/auth/AuthorizationForm";
-
-// export function AuthPage() {
-//   return (
-//     <>
-//       <AuthorizationForm />
-//     </>
-//   );
-// }
