+++ client/src/pages/LandingPage/index.ts
@@ -1,0 +1,1 @@
+export { LandingPage } from "./ui/LandingPage.tsx";
