--- client/src/pages/RegPage/index.ts
@@ -1,1 +1,0 @@
-// export { RegPage } from './ui/RegPage';
