--- client/src/pages/RegPage/ui/RegPage.tsx
@@ -1,9 +1,0 @@
-// import { RegistrationForm } from "@/features/auth/RegistrationForm";
-
-// export function RegPage() {
-//   return (
-//     <>
-//       <RegistrationForm />
-//     </>
-//   );
-// }
