+++ client/src/pages/index.ts
@@ -1,5 +1,3 @@
-// export { AuthPage } from "./AuthPage";
 export { HomePage } from "./HomePage";
-// export { RegPage } from "./RegPage";
 export { NotFound } from "./NotFound";
-export { LoginPage } from "./LoginPage";
+export { LandingPage } from "./LandingPage";
