+++ client/src/shared/ui/Loader/Loader.tsx
@@ -1,4 +1,4 @@
-import './Loader.css';
+// import './Loader.css';
 
 export function Loader(): JSX.Element {
     return (
