--- client/src/widgets/Nav/ui/Nav.css
@@ -1,3 +1,0 @@
-.custom-navbar {
-    background-color: #2c3e50;
-  }
