+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-hook-form": "^7.54.0"
+  }
+}
