+++ server/routes/api.routes.js
@@ -4,7 +4,7 @@ const apiTicketsRouter = require("./api.ticket.routes");
 const apiCommentsRouter = require("./api.comment.routes");
 
 
-router.use("/auth", apiAuthRouter);
+router.use("/auth", apiUserRouter);
 router.use("/tickets", apiTicketsRouter)
 router.use("/comments", apiCommentsRouter)
 module.exports = router;
